Create string variable
C# is a strongly typed language.  This means that the data-type must also
be declared when creating a variable.

I created a variable with the message Hello World!  Then reassigned the
variable to World!.  I wanted to practice concatenating a string and a
variable and printing to the console.

This is all practice getting familiar with C#.  I will finish this
repository with a simple application demonstrating basic C# proficiency.