Project Build Succeeded
Add a connection string to the appsettings.json file to connect to the
database.

I also installed required packages to run the project.  These packages
include SQLite, CodeGeneration.Design for scaffolding, and
EntityFrameworkCore.

I then add the using statements in Startup.cs file to use the model I
created and the Microsoft.EntityFrameworkCore.

Finally, the database context was registered with the dependency
injection container in Startup.ConfigureServerices.  Build was run to
test the changes and succeeded.