Fix 404 error
When I started server using dotnet run and naviagted to where my
database should display, there was a 404 error.

The problem is when I ran the command to create the database CRUD pages,
I didn't put a / to denote a nested folder in the directory, so it was
created as Pages/Movies.

Once created the folder Movies and nested the .cshtml files inside of
the folder, I was able to navigate to the requested page.  I also tested
CRUD capabilities and everything check out.

I am now moving on to the next part of this tutorial.